data/crawled_pages/https___firefox_source_docs_mozilla_org_testing_chrome_tests_index_html.html

<!doctype html>
<html class="writer-html5" lang="en" data-content_root="../../">
 <head>
  <meta charset="utf-8">
  <meta name="viewport" content="width=device-width, initial-scale=1">
  <meta name="viewport" content="width=device-width, initial-scale=1.0">
  <title>Chrome Tests — Firefox Source Docs  documentation</title>
  <link rel="stylesheet" type="text/css" href="../../_static/pygments.css?v=fa44fd50">
  <link rel="stylesheet" type="text/css" href="../../_static/css/theme.css?v=19f00094">
  <link rel="stylesheet" type="text/css" href="../../_static/graphviz.css?v=fd3f3429">
  <link rel="stylesheet" type="text/css" href="../../_static/copybutton.css?v=76b2166b">
  <link rel="stylesheet" type="text/css" href="../../_static/custom_theme.css?v=dc5ed815">
  <link rel="stylesheet" type="text/css" href="../../_static/design-style.1e8bd061cd6da7fc9cf755528e8ffc24.min.css?v=0a3b3ea7">
  <link rel="shortcut icon" href="../../_static/firefox.ico"><!--[if lt IE 9]>
    <script src="../../_static/js/html5shiv.min.js"></script>
  <![endif]-->
  <script src="../../_static/jquery.js?v=5d32c60e"></script>
  <script src="../../_static/_sphinx_javascript_frameworks_compat.js?v=2cd50e6c"></script>
  <script src="../../_static/documentation_options.js?v=5929fcd5"></script>
  <script src="../../_static/doctools.js?v=9a2dae69"></script>
  <script src="../../_static/sphinx_highlight.js?v=dc90522c"></script>
  <script src="../../_static/clipboard.min.js?v=a7894cd8"></script>
  <script src="../../_static/copybutton.js?v=30646c52"></script>
  <script src="../../_static/design-tabs.js?v=36754332"></script>
  <script src="../../_static/js/theme.js"></script>
  <link rel="index" title="Index" href="../../genindex.html">
  <link rel="search" title="Search" href="../../search.html">
  <link rel="next" title="Marionette" href="../marionette/index.html">
  <link rel="prev" title="Browser chrome mochitests" href="../browser-chrome/index.html">
 </head>
 <body class="wy-body-for-nav">
  <div class="wy-grid-for-nav">
   <nav data-toggle="wy-nav-shift" class="wy-nav-side">
    <div class="wy-side-scroll">
     <div class="wy-side-nav-search">
      <a href="../../index.html" class="icon icon-home"> Firefox Source Docs <img src="../../_static/firefox-wordmark.svg" class="logo" alt="Logo"> </a><!-- -- This code is governed by the BSD license ---->
      <div>
       <h3>Quick search</h3>
       <script>
    (function () {
      var cx = "dd12886298f75dbef";
      var gcse = document.createElement("script");
      gcse.async = true;
      gcse.src = "https://cse.google.com/cse.js?cx=" + cx;
      var s = document.getElementsByTagName("script")[0];
      s.parentNode.insertBefore(gcse, s);
    })();
  </script><gcse:search></gcse:search>
      </div>
     </div>
     <div class="wy-menu wy-menu-vertical" data-spy="affix" role="navigation" aria-label="Navigation menu">
      <p class="caption" role="heading"><span class="caption-text">Overview</span></p>
      <ul>
       <li class="toctree-l1"><a class="reference internal" href="../../glossary/index.html">A Glossary of Common Terms</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../overview/index.html">A Quick Guide to Mozilla Applications</a></li>
      </ul>
      <p class="caption" role="heading"><span class="caption-text">Getting Started</span></p>
      <ul>
       <li class="toctree-l1"><a class="reference internal" href="../../setup/index.html">Getting Set Up To Work On The Firefox Codebase</a></li>
      </ul>
      <p class="caption" role="heading"><span class="caption-text">Working On Firefox</span></p>
      <ul>
       <li class="toctree-l1"><a class="reference internal" href="../../contributing/index.html">Working on Firefox</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../bug-mgmt/index.html">Bug Handling</a></li>
      </ul>
      <p class="caption" role="heading"><span class="caption-text">Firefox User Guide</span></p>
      <ul>
       <li class="toctree-l1"><a class="reference internal" href="../../devtools-user/index.html">Firefox DevTools User Docs</a></li>
      </ul>
      <p class="caption" role="heading"><span class="caption-text">Source Code Documentation</span></p>
      <ul>
       <li class="toctree-l1"><a class="reference internal" href="../../mots/index.html">Governance</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../browser/index.html">Firefox Front-end</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../dom/index.html">DOM</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../editor/index.html">Editor</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../layout/index.html">Style system (CSS) &amp; Layout</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../gfx/index.html">Graphics</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../ipc/index.html">Processes, Threads and IPC</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../devtools/index.html">Firefox DevTools Contributor Docs</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../toolkit/index.html">Toolkit</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../js/index.html">SpiderMonkey</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../jsloader/index.html">JS Loader</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../mobile/android/geckoview/index.html">GeckoView</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../mobile/android/fenix/index.html">Fenix</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../mobile/android/focus-android/index.html">Focus for Android</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../dom/bindings/webidl/index.html">WebIDL</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../modules/libpref/index.html">libpref</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../networking/index.html">Networking</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../remote/index.html">Remote Protocols</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../services/index.html">Services</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../permissions/index.html">Permissions</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../uriloader/index.html">File Handling</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../widget/cocoa/index.html">Firefox on macOS</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../widget/windows/index.html">Firefox on Windows</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../toolkit/components/ml/index.html">Firefox AI Runtime</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../accessible/index.html">Accessibility</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../media/index.html">Media Playback</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../code-quality/index.html">Code quality</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../writing-rust-code/index.html">Writing Rust Code</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../rust-components/index.html">Rust Components</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../tools/profiler/index.html">Gecko Profiler</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../performance/index.html">Performance</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../storage/index.html">Database bindings (SQLite, KV, …)</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../xpcom/index.html">XPCOM</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../nspr/index.html">NSPR</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../security/nss/index.html">Network Security Services (NSS)</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../content-security/index.html">Web Security Checks in Gecko</a></li>
      </ul>
      <p class="caption" role="heading"><span class="caption-text">The Firefox Build System</span></p>
      <ul>
       <li class="toctree-l1"><a class="reference internal" href="../../mach/index.html">Mach</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../tools/try/index.html">Pushing to Try</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../build/buildsystem/index.html">Build System</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../taskcluster/index.html">Firefox CI and Taskgraph</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../tools/moztreedocs/index.html">Managing Documentation</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../mozbuild/vendor/index.html">Vendoring Third Party Components</a></li>
      </ul>
      <p class="caption" role="heading"><span class="caption-text">Testing &amp; Test Infrastructure</span></p>
      <ul class="current">
       <li class="toctree-l1"><a class="reference internal" href="../automated-testing/index.html">Automated Testing</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../treeherder-try/index.html">Understanding Treeherder Results</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../sheriffed-intermittents/index.html">Sheriffed intermittent failures</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../tests-for-new-config/index.html">Turning on Firefox tests for a new configuration</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../intermittent/index.html">Avoiding intermittent tests</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../testing-policy/index.html">Testing Policy</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../ci-configs/index.html">Configuration Changes</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../browser-chrome/index.html">Browser chrome mochitests</a></li>
       <li class="toctree-l1 current"><a class="current reference internal" href="#">Chrome Tests</a>
        <ul>
         <li class="toctree-l2"><a class="reference internal" href="#disclaimer"><strong>DISCLAIMER</strong></a></li>
         <li class="toctree-l2"><a class="reference internal" href="#introduction">Introduction</a></li>
         <li class="toctree-l2"><a class="reference internal" href="#running-the-chrome-tests">Running the chrome tests</a></li>
         <li class="toctree-l2"><a class="reference internal" href="#writing-chrome-tests">Writing chrome tests</a></li>
         <li class="toctree-l2"><a class="reference internal" href="#adding-a-new-chrome-test-to-the-tree">Adding a new chrome test to the tree</a></li>
        </ul></li>
       <li class="toctree-l1"><a class="reference internal" href="../marionette/index.html">Marionette</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../geckodriver/index.html">geckodriver</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../test-verification/index.html">Test Verification</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../webrender/index.html">WebRender Tests</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../mochitest-plain/index.html">Mochitest</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../xpcshell/index.html">XPCShell tests</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../tps/index.html">TPS</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../web-platform/index.html">web-platform-tests</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../gtest/index.html">GTest</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../tools/fuzzing/index.html">Fuzzing</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../tools/sanitizer/index.html">Sanitizer</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../perfdocs/index.html">Performance Testing</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../tools/code-coverage/index.html">Code coverage</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../testing-rust-code/index.html">Testing &amp; Debugging Rust Code</a></li>
      </ul>
      <p class="caption" role="heading"><span class="caption-text">Releases &amp; Updates</span></p>
      <ul>
       <li class="toctree-l1"><a class="reference internal" href="../../update-infrastructure/index.html">Mozilla Update Infrastructure</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../update-infrastructure/index.html#watershed-updates">Watershed Updates</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../update-infrastructure/index.html#desupport-updates">Desupport Updates</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../tools/update-verify/index.html">Update Verify</a></li>
      </ul>
      <p class="caption" role="heading"><span class="caption-text">Localization &amp; Internationalization</span></p>
      <ul>
       <li class="toctree-l1"><a class="reference internal" href="../../intl/index.html">Internationalization</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../l10n/index.html">Localization</a></li>
      </ul>
      <p class="caption" role="heading"><span class="caption-text">Firefox and Python</span></p>
      <ul>
       <li class="toctree-l1"><a class="reference internal" href="../../mozbase/index.html">mozbase</a></li>
       <li class="toctree-l1"><a class="reference internal" href="../../python/index.html">Using third-party Python packages</a></li>
      </ul>
      <p class="caption" role="heading"><span class="caption-text">Metrics Collected in Firefox</span></p>
      <ul>
       <li class="toctree-l1"><a class="reference internal" href="../../metrics/index.html">Metrics</a></li>
      </ul>
     </div>
    </div>
   </nav>
   <section data-toggle="wy-nav-shift" class="wy-nav-content-wrap">
    <nav class="wy-nav-top" aria-label="Mobile navigation menu">
     <i data-toggle="wy-nav-top" class="fa fa-bars"></i> <a href="../../index.html">Firefox Source Docs</a>
    </nav>
    <div class="wy-nav-content">
     <div class="rst-content">
      <!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this file,
   - You can obtain one at http://mozilla.org/MPL/2.0/.  -->
      <div role="navigation" aria-label="Page navigation">
       <ul class="wy-breadcrumbs">
        <li><a href="../../index.html" class="icon icon-home" aria-label="Home"></a></li>
        <li class="breadcrumb-item active">Chrome Tests</li>
        <li class="wy-breadcrumbs-aside"><a href="https://bugzilla.mozilla.org/enter_bug.cgi?product=Developer+Infrastructure&amp;component=Firefox+Source+Docs%3A+Content&amp;short_desc=Documentation+issue+on+testing/chrome-tests/index&amp;comment=URL+=+https://firefox-source-docs.mozilla.org/testing/chrome-tests/index.html&amp;bug_file_loc=https://firefox-source-docs.mozilla.org/testing/chrome-tests/index.html" rel="nofollow">Report an issue</a> / <a href="../../_sources/testing/chrome-tests/index.rst.txt" rel="nofollow"> View page source</a></li>
       </ul>
       <hr>
      </div>
      <div role="main" class="document" itemscope itemtype="http://schema.org/Article">
       <div itemprop="articleBody">
        <section id="chrome-tests">
         <h1>Chrome Tests<a class="headerlink" href="#chrome-tests" title="Link to this heading">?</a></h1>
         <section id="disclaimer">
          <span id="id1"></span>
          <h2><strong>DISCLAIMER</strong><a class="headerlink" href="#disclaimer" title="Link to this heading">?</a></h2>
          <p><strong>NOTE: Please use this document as a reference for existing chrome tests as you do not want to create new chrome tests. If you’re trying to test privileged browser code, write a browser mochitest instead; if you are testing web platform code, use a wpt test, or a “plain” mochitest if you are unable to use a wpt test.</strong></p>
         </section>
         <section id="introduction">
          <span id="id2"></span>
          <h2>Introduction<a class="headerlink" href="#introduction" title="Link to this heading">?</a></h2>
          <p>A chrome test is similar but not equivalent to a Mochitest running with chrome privileges.</p>
          <p>The chrome test suite is an automated testing framework designed to allow testing of application chrome windows using JavaScript. It allows you to run JavaScript code in the non-electroysis (e10s) content area with chrome privileges, instead of directly in the browser window (as browser tests do instead). These tests reports results using the same functions as the Mochitest test framework. The chrome test suite depends on runtests.py from the Mochitest framework.</p>
         </section>
         <section id="running-the-chrome-tests">
          <span id="id3"></span>
          <h2>Running the chrome tests<a class="headerlink" href="#running-the-chrome-tests" title="Link to this heading">?</a></h2>
          <p>To run chrome tests, you need to <a class="reference external" href="/setup">build Firefox</a> with your changes and find the test or test manifest you want to run.</p>
          <p>For example, to run all chrome tests under <cite>toolkit/content</cite>, run the following command:</p>
          <div class="highlight-default notranslate">
           <div class="highlight">
            <pre><span></span><span class="o">./</span><span class="n">mach</span> <span class="n">test</span> <span class="n">toolkit</span><span class="o">/</span><span class="n">content</span><span class="o">/</span><span class="n">test</span><span class="o">/</span><span class="n">chrome</span><span class="o">/</span><span class="n">chrome</span><span class="o">.</span><span class="n">ini</span>
</pre>
           </div>
          </div>
          <p>To run a single test, just pass the path to the test into mach:</p>
          <div class="highlight-default notranslate">
           <div class="highlight">
            <pre><span></span><span class="o">./</span><span class="n">mach</span> <span class="n">test</span> <span class="n">toolkit</span><span class="o">/</span><span class="n">content</span><span class="o">/</span><span class="n">tests</span><span class="o">/</span><span class="n">chrome</span><span class="o">/</span><span class="n">test_largemenu</span><span class="o">.</span><span class="n">html</span>
</pre>
           </div>
          </div>
          <p>You can also pass the path to a directory containing many tests. Run <cite>./mach test –help</cite> for full documentation.</p>
         </section>
         <section id="writing-chrome-tests">
          <span id="id4"></span>
          <h2>Writing chrome tests<a class="headerlink" href="#writing-chrome-tests" title="Link to this heading">?</a></h2>
          <p>A chrome tests is similar but not equivalent to a Mochitest running with chrome privileges, i.e. code and UI are referenced by <code class="docutils literal notranslate"><span class="pre">chrome://</span></code> URIs. A basic XHTML test file could look like this:</p>
          <div class="highlight-xml notranslate">
           <div class="highlight">
            <pre><span></span><span class="cp">&lt;?xml version="1.0"?&gt;</span>
<span class="cp">&lt;?xml-stylesheet href="chrome://global/skin" type="text/css"?&gt;</span>
<span class="cp">&lt;?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css" type="text/css"?&gt;</span>

<span class="nt">&lt;window</span><span class="w"> </span><span class="na">title=</span><span class="s">"Demo Test"</span>
<span class="w">        </span><span class="na">xmlns=</span><span class="s">"http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"</span><span class="nt">&gt;</span>
<span class="w">  </span><span class="nt">&lt;title&gt;</span>Demo<span class="w"> </span>Test<span class="nt">&lt;/title&gt;</span>

<span class="w">  </span><span class="nt">&lt;script</span><span class="w"> </span><span class="na">type=</span><span class="s">"application/javascript"</span>
<span class="w">          </span><span class="na">src=</span><span class="s">"chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"</span><span class="nt">/&gt;</span>

<span class="w">  </span><span class="nt">&lt;script</span><span class="w"> </span><span class="na">type=</span><span class="s">"application/javascript"</span><span class="nt">&gt;</span>
<span class="w">  </span><span class="cp">&lt;![CDATA[</span>
<span class="cp">    add_task(async function runTest() {</span>
<span class="cp">       ok  (true ==  1, "this passes");</span>
<span class="cp">      todo(true === 1, "this fails");</span>
<span class="cp">    });</span>
<span class="cp">  ]]&gt;</span>
<span class="w">  </span><span class="nt">&lt;/script&gt;</span>

<span class="w">  </span><span class="nt">&lt;body</span><span class="w"> </span><span class="na">xmlns=</span><span class="s">"http://www.w3.org/1999/xhtml"</span><span class="nt">&gt;</span>
<span class="w">    </span><span class="nt">&lt;p</span><span class="w"> </span><span class="na">id=</span><span class="s">"display"</span><span class="nt">&gt;&lt;/p&gt;</span>
<span class="w">    </span><span class="nt">&lt;div</span><span class="w"> </span><span class="na">id=</span><span class="s">"content"</span><span class="w"> </span><span class="na">style=</span><span class="s">"display: none"</span><span class="nt">&gt;&lt;/div&gt;</span>
<span class="w">    </span><span class="nt">&lt;pre</span><span class="w"> </span><span class="na">id=</span><span class="s">"test"</span><span class="nt">&gt;&lt;/pre&gt;</span>
<span class="w">  </span><span class="nt">&lt;/body&gt;</span>
<span class="nt">&lt;/window&gt;</span>
</pre>
           </div>
          </div>
          <p>The comparison functions are identical to those supported by Mochitests, see how the comparison functions work in the Mochitest documentation for more details. The <a class="reference external" href="https://searchfox.org/mozilla-central/source/testing/mochitest/tests/SimpleTest/EventUtils.js">EventUtils helper functions</a> are available on the “EventUtils” object defined in the global scope.</p>
          <p>The test suite also supports asynchronous tests. To use these asynchronous tests, please use the <a class="reference external" href="https://searchfox.org/mozilla-central/source/testing/mochitest/tests/SimpleTest/SimpleTest.js#2025">add_task()</a> functionality.</p>
          <p>Any exceptions thrown while running a test will be caught and reported in the test output as a failure. Exceptions thrown outside of the test’s context (e.g. in a timeout, event handler, etc) will not be caught, but will result in a timed out test.</p>
          <p>The test file name must be prefixed with <code class="docutils literal notranslate"><span class="pre">test_</span></code>, and must have a file extension of <code class="docutils literal notranslate"><span class="pre">.xhtml</span></code>. Files that don’t match this pattern will be ignored by the test harness, but you still can include them. For example, a XUL window file opened by your test_demo.xhtml via openDialog should be named window_demo.xhtml. Putting the bug number in the file name is recommended if your test verifies a bugfix, e.g. “test_bug123456.xhtml”.</p>
          <p>Helper files can be included, for example, from <code class="docutils literal notranslate"><span class="pre">https://example.com/chrome/dom/workers/test/serviceworkers/serviceworkermanager_iframe.html</span></code>.</p>
         </section>
         <section id="adding-a-new-chrome-test-to-the-tree">
          <span id="id5"></span>
          <h2>Adding a new chrome test to the tree<a class="headerlink" href="#adding-a-new-chrome-test-to-the-tree" title="Link to this heading">?</a></h2>
          <p>To add a new chrome test to the tree, please use <cite>./mach test addtest the_test_directory/the_test_you_want_to_create.xhtml</cite>. For more information about <cite>addtest</cite>, please run <cite>./mach test addtest –help</cite>.</p>
         </section>
        </section>
       </div>
      </div>
      <footer>
       <div class="rst-footer-buttons" role="navigation" aria-label="Footer">
        <a href="../browser-chrome/index.html" class="btn btn-neutral float-left" title="Browser chrome mochitests" accesskey="p" rel="prev"><span class="fa fa-arrow-circle-left" aria-hidden="true"></span> Previous</a> <a href="../marionette/index.html" class="btn btn-neutral float-right" title="Marionette" accesskey="n" rel="next">Next <span class="fa fa-arrow-circle-right" aria-hidden="true"></span></a>
       </div>
       <hr>
       <div role="contentinfo">
        <p></p>
       </div> Built with <a href="https://www.sphinx-doc.org/">Sphinx</a> using a <a href="https://github.com/readthedocs/sphinx_rtd_theme">theme</a> provided by <a href="https://readthedocs.org">Read the Docs</a>.
      </footer>
     </div>
    </div>
   </section>
  </div>
  <script>
      jQuery(function () {
          SphinxRtdTheme.Navigation.enable(true);
      });
  </script>
 </body>
</html>
